fix: #1058 ベーストークン直使用修正 + fetchUmamiData をサービス層に抽出
- changeColorClass: --color-success/danger → --color-action-success/danger (セマンティックトークン)
- <style>: --color-neutral-700/900/100 → --color-text-primary/text/border-light
- fetchUmamiData() を +page.server.ts → $lib/server/services/umami-service.ts に移動
- AnalyticsData 型を re-export してフロントのインポートパスを維持
